Added random initial data generator.
Added a function to generator a random function with periodic boundary conditions. The smoothness parameter, reg, in this function needs to be taken relatively large, say at least around 2-3. If 'reg' is too small, the randomly generator initial tends to be very jagged, making the finite difference method not applicable.